Still working on models.py It seems that supervised sessions loads the last trained model for initialization causing unwanted errors

+++ DNA_reader.py
@@ -69,7 +69,7 @@ def DNA_producer(raw_data, config, name=None):
         num_batches = data_size // batch_size
         data = raw_data[0: batch_size*num_batches]
         labels = labels[0: batch_size*num_batches]
-        labels = tf.convert_to_tensor(labels, name="encoder_labels", dtype=tf.int64)
+        labels = tf.convert_to_tensor(labels, name="encoder_labels", dtype=tf.int32)
         # vocab_size -1 is 7 in our case and means PAD symbol
         encoder_input_full = np.ones([data_size, num_steps_encoder]) *(vocab_size-1)
         decoder_input_full = np.ones([data_size, num_steps_decoder]) *(vocab_size-1)
